+++ src/app/(dashboard)/dashboard/customer/layout.js
@@ -3,6 +3,7 @@ import { AppSidebar } from '@/app/components/Pages/DASHBOARD/user/app-sidebar';
 import { Toaster } from '@/components/ui/toaster';
 
 export default function UserLayout({ children }) {
+  
   return (
     <SidebarProvider>
       <AppSidebar />