+++ src/app/(frontend)/blogs/[slug]/page.js
@@ -4,6 +4,12 @@ import ContentSection from '@/app/components/Pages/FRONT_END/singleblog/ContentS
 import GuideSection from '@/app/components/Pages/FRONT_END/Global/GuideSection';
 import { fakeData } from '@/app/Data/ShopData';
 
+export async function generateMetadata() {
+  return {
+    title: 'Single Blog Page',
+    description: 'Description of the Single Blog',
+  };
+}
 const SingleBlogPage = async ({ params }) => {
   const { slug } = await params;