+++ src/app/(frontend)/package/[package]/page.js
@@ -3,24 +3,27 @@ import { log } from '@/lib/utils';
 import { notFound } from 'next/navigation';
 import { publicApi } from '@/lib/axiosInstance';
 import BannerSection from '@/app/components/Pages/FRONT_END/singleproduct/BannerSection';
-import { TabSectionIterenary, TabSectionPackage } from '@/app/components/Pages/FRONT_END/singleproduct/TabSection';
+import { TabSectionPackage } from '@/app/components/Pages/FRONT_END/singleproduct/TabSection';
+import { getSinglePackage } from '@/lib/services/package';
+import { isEmpty } from 'lodash';
 
-// Fetch package data
-async function getPackageData(packagee) {
-  try {
-    const response = await publicApi.get(`/api/packages/${packagee}`, {
-      headers: { Accept: 'application/json' },
-    });
+export async function generateMetadata({ params }) {
+  const { package: pack } = await params;
+  const { data: packageData = [] } = await getSinglePackage(pack);
 
-    return response.data;
-  } catch (error) {
-    return []; // Return null instead of an empty array for clarity
+  if (!isEmpty(packageData)) {
+    const { name, description } = packageData;
+
+    return {
+      title: name,
+      description,
+    };
   }
 }
 
 export default async function PackagePage({ params }) {
   const { package: pack } = await params;
-  const { data: packageData = [] } = await getPackageData(pack);
+  const { data: packageData = [] } = await getSinglePackage(pack);
 
   // Check if packageData is null or empty
   if (!packageData || packageData.length === 0) {