+++ src/lib/services/package.js
@@ -2,6 +2,25 @@
 import { authApi, publicApi } from '../axiosInstance';
 import { log } from '../utils';
 
+
+/**
+ * Get Single Package on Public side
+ * @param {Number} id
+ * @returns []
+ */
+export async function getSinglePackage(packagee) {
+  try {
+    const response = await publicApi.get(`/api/packages/${packagee}`, {
+      headers: { Accept: 'application/json' },
+    });
+
+    return response.data;
+  } catch (error) {
+    return []; // Return null instead of an empty array for clarity
+  }
+}
+
+
 /**
  * Get Single Package on Admin side
  * @param {Number} id